+++ packages/client/window/src/windows/Popup.ts
@@ -1,3 +1,4 @@
+import { CreatePopupStrategy } from "@/utils/popupStrategy";
 import { urlToOrigin } from "@/utils/urlToOrigin";
 
 import { EventMap } from "../EventEmitter";
@@ -7,6 +8,7 @@ import { HandshakeParent } from "../handshake/Parent";
 export interface PopupWindowOptions {
     width: number;
     height: number;
+    crossOrigin?: boolean;
 }
 
 export class PopupWindow<IncomingEvents extends EventMap, OutgoingEvents extends EventMap> extends HandshakeParent<
@@ -25,8 +27,9 @@ export class PopupWindow<IncomingEvents extends EventMap, OutgoingEvents extends
         url: string,
         options: PopupWindowOptions & EventEmitterWithHandshakeOptions<IncomingEvents, OutgoingEvents>
     ) {
+        const popup = await createPopup(url, options);
         return new PopupWindow<IncomingEvents, OutgoingEvents>(
-            await createPopup(url, options),
+            popup,
             options?.targetOrigin || urlToOrigin(url),
             options
         );
@@ -34,7 +37,11 @@ export class PopupWindow<IncomingEvents extends EventMap, OutgoingEvents extends
 }
 
 async function createPopup(url: string, options: PopupWindowOptions): Promise<Window> {
-    const _window = window.open(url, "popupWindow", createPopupString(options.width, options.height));
+    const _window = window.open(
+        url,
+        "popupWindow",
+        createPopupString(options.width, options.height, options?.crossOrigin || false)
+    );
     if (!_window) {
         throw new Error("Failed to open popup window");
     }
@@ -45,43 +52,25 @@ async function createPopup(url: string, options: PopupWindowOptions): Promise<Wi
     });
 }
 
-function createPopupString(width: number, height: number): string {
-    function getLeft() {
-        try {
-            return window?.top != null
-                ? window.top.outerWidth / 2 + window.top.screenX - width / 2
-                : window.outerWidth / 2 + window.screenX - width / 2;
-        } catch (e) {
-            console.error(e);
-        }
-
-        return window.outerWidth / 2 + window.screenX - width / 2;
-    }
-
-    function getTop() {
-        try {
-            return window?.top != null
-                ? window.top.outerHeight / 2 + window.top.screenY - height / 2
-                : window.outerHeight / 2 + window.screenY - height / 2;
-        } catch (e) {
-            console.error(e);
-        }
-
-        return window.outerHeight / 2 + window.screenY - height / 2;
-    }
-
-    function getChromeVersion() {
-        const raw = navigator.userAgent.match(/Chrom(e|ium)\/([0-9]+)\./);
-        return raw ? parseInt(raw[2]) : null;
-    }
-    function isFirefox() {
-        return navigator.userAgent.toLowerCase().indexOf("firefox") > -1;
-    }
+function createPopupString(width: number, height: number, crossOrigin: boolean): string {
+    const createPopupStrategy = new CreatePopupStrategy(crossOrigin);
 
     // In newer versions of chrome (>99) you need to add the `popup=true` for the new window to actually open in a popup
     const chromeVersion = getChromeVersion();
     const chromeVersionGreaterThan99 = chromeVersion != null && chromeVersion > 99;
     const popupStringBase = isFirefox() || chromeVersionGreaterThan99 ? "popup=true," : "";
 
-    return `${popupStringBase}height=${height},width=${width},left=${getLeft()},top=${getTop()},resizable=yes,scrollbars=yes,toolbar=yes,menubar=true,location=no,directories=no,status=yes`;
+    return `${popupStringBase}height=${height},width=${width},left=${createPopupStrategy.getLeft(
+        width
+    )},top=${createPopupStrategy.getTop(
+        height
+    )},resizable=yes,scrollbars=yes,toolbar=yes,menubar=true,location=no,directories=no,status=yes`;
+}
+
+function getChromeVersion() {
+    const raw = navigator.userAgent.match(/Chrom(e|ium)\/([0-9]+)\./);
+    return raw ? parseInt(raw[2]) : null;
+}
+function isFirefox() {
+    return navigator.userAgent.toLowerCase().indexOf("firefox") > -1;
 }