+++ packages/client/ui/react-ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-react-ui
 
+## 1.2.2
+
+### Patch Changes
+
+-   Updated dependencies [a9b57fa]
+    -   @crossmint/common-sdk-base@0.1.0
+    -   @crossmint/client-sdk-base@1.2.1
+
 ## 1.2.1
 
 ### Patch Changes