+++ packages/client/ui/vanilla-ui/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @crossmint/client-sdk-vanilla-ui
 
+## 1.1.11
+
+### Patch Changes
+
+-   Updated dependencies [a9b57fa]
+    -   @crossmint/common-sdk-base@0.1.0
+    -   @crossmint/client-sdk-base@1.2.1
+
 ## 1.1.10
 
 ### Patch Changes