+++ packages/client/window/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @crossmint/client-sdk-window
 
+## 0.0.8
+
+### Patch Changes
+
+-   dbca75f: Releasing Window SDK to be used in cross origin communication
+
 ## 0.0.7
 
 ### Patch Changes